Make amp.php compatibility with PHP 5.2
* Move functions to amp-helper-functions.php.
* Add amp_bootstrap_plugin() function to contain former top-level calls
